TQS-77 Create Endpoints to list orders, get a specific order of a rider

+++ projDelivera/backend/delivera/src/main/java/ua/tqs/delivera/models/OrderProfit.java
@@ -29,6 +29,6 @@ public class OrderProfit {
   private Order order;
   
   @Column(name = "order_price")
-  private long orderPrice;
+  private Double orderPrice;
   
 }
